// 2019/website/src/widgets/common/interfaces.d.ts
export type AriaPropertyObject = {
	[key: string]: string;
};

export interface CustomAriaProperties {
	aria?: AriaPropertyObject;
}

export interface Focus extends FocusEvent {
	readonly value?: string;
}
export interface Mouse extends MouseEvent {
	readonly previousScreenX: number;
	readonly previousScreenY: number;
	readonly movementX: number;
	readonly movementY: number;
	readonly relatedTarget: any;
	readonly targetKey: string | null;
}
export interface Input extends UIEvent {
	readonly value: string;
	readonly data: string;
	readonly dataTransfer: DataTransfer;
	readonly inputType: 'insertText' | 'insertCompositionText' | 'insertFromComposition' |
	'insertFromPaste' | 'insertFromDrop' | 'insertReplacementText' | 'insertFromYank'|
	'formatSetBlockTextDirection' | 'formatSetInlineTextDirection' | 'formatBackColor'|
	'formatFontColor' | 'formatFontName' | 'insertLink' | 'historyUndo' | 'historyRedo';
  readonly isComposing: boolean;
}
export interface Toggle extends MouseEvent {
	readonly value: string;
	readonly checked: boolean;
}
export interface Keyboard extends KeyboardEvent {
	readonly which: number;
	readonly key: Keys;
	readonly isStart: boolean;
	readonly isEnd: boolean;
}
// ...
export interface LabeledProperties {
	labelAfter?: boolean;
	labelHidden?: boolean;
	labelStatic?: boolean;
	label?: DNode;
	helperText?: string;
}
// ...
export interface KeyEventProperties {
	onKeyDown?(event: Keyboard, preventDefault: () => void): void;
	onKeyPress?(event: Keyboard, preventDefault: () => void): void;
	onKeyUp?(event: Keyboard, preventDefault: () => void): void;
}

/**
 * @type PointerEventProperties
 * @property onMouseDown    Called on the input's mousedown event
 * @property onMouseUp      Called on the input's mouseup event
 * @property onTouchStart   Called on the input's touchstart event
 * @property onTouchEnd     Called on the input's touchend event
 * @property onTouchCancel  Called on the input's touchcancel event
 */
export interface PointerEventProperties {
	onMouseDown?(event: MouseEvent): void;
	onMouseUp?(event: MouseEvent): void;
	onTouchStart?(event: TouchEvent): void;
	onTouchEnd?(event: TouchEvent): void;
	onTouchCancel?(event: TouchEvent): void;
}
// ...
export interface RedaktorProperties extends ThemedProperties {
  required?: boolean;
  disabled?: boolean;
  readOnly?: boolean;
  invalid?: boolean;
  responsive?: boolean;
	size?: Sizes | undefined;
	schema?: any; // MaterialSchema | keyof typeof MaterialSchema; // TODO
  filled?: boolean;
  outlined?: boolean;
  shaped?: boolean;
}